feat: 사용하지 않는 타입 및 메서드 삭제
- Holder타입 및 checkStatusCode메서드
